Server-modifiable-Opacity
This adds two lines to the config file, one for line opacity and one for fill opacity. These accept values from 0 to 255 and correspondingly set the opacity of the line and fill.

Setting the value to 0 will be effectively invisible, setting it to 255 will be very solid and obscure pretty much anything beneath it.